improve ascom check

+++ tests/unit_tests/logic/camera/test_cameraAscom.py
@@ -108,7 +108,7 @@ def test_workerExpose_1(function):
     with mock.patch.object(function.parent, "sendDownloadMode"):
         with mock.patch.object(function, "setAscomProperty"):
             with mock.patch.object(function.parent, "waitExposed"):
-                with mock.patch.object(function, "getAscomProperty", return_values=np.ones(10, 10)):
+                with mock.patch.object(function, "getAscomProperty", return_values=np.ones((10, 10), dtype=np.float32)):
                     with mock.patch.object(function.parent, "writeImageFitsHeader"):
                         with mock.patch.object(fits.HDUList, "writeto"):
                             function.workerExpose()
